add footer class
